refactor(core/dataTable): revisions draft overview dataTableType
