+++ EMS/core-bundle/src/Resources/config/controllers.xml
@@ -297,11 +297,11 @@
         </service>
         <service id="EMS\CoreBundle\Controller\Revision\EditController" public="true">
             <argument type="service" id="ems.service.data" />
-            <argument type="service" id="ems_core.core_revision.draft_in_progress" />
             <argument type="service" id="logger" />
             <argument type="service" id="ems.service.publish" />
             <argument type="service" id="ems.service.revision" />
             <argument type="service" id="translator" />
+            <argument type="service" id="emsco.data_table.factory" />
             <call method="setContainer"/>
             <tag name="container.service_subscriber"/>
             <tag name="controller.service_arguments"/>
